Rename EmptyStates to constants
