+++ src/SmartComponents/Inventory/applications/vulnerabilities/constants.js
